feat: fix the tests
