+++ whois-api/src/main/java/net/ripe/db/whois/api/rdap/WhoisRdapService.java
@@ -370,10 +370,6 @@ public class WhoisRdapService {
 
         final RpslObject resultObject = rpslIterator.next();
 
-        if (isIANABlock(resultObject)){
-            throw new RdapException("404 Not Found", "Requested object not found", HttpStatus.NOT_FOUND_404);
-        }
-
         if (rpslIterator.hasNext()) {
             throw new RdapException("500 Internal Error", "Unexpected result size: " + Iterators.size(rpslIterator),
                     HttpStatus.INTERNAL_SERVER_ERROR_500);
@@ -386,10 +382,6 @@ public class WhoisRdapService {
                         abuseCFinder.getAbuseContact(resultObject).orElse(null)))
                 .header(CONTENT_TYPE, CONTENT_TYPE_RDAP_JSON)
                 .build();
-    }
-
-    private boolean isIANABlock(final RpslObject rpslObject) {
-        return rpslObject.getKey().toString().equals("::/0") || rpslObject.getKey().toString().equals("0.0.0.0 - 255.255.255.255");
     }
 
     private Response redirect(final String requestPath, final Query query) {
